Create sidebar collapse/expand feature
The sidebar now can collapse when the icon button is pressed - no animations have been added yet and the app layout is not responsive to the closed sidebar. SidebarActions have been renamed to SidebarToolbar. The sidebar now renders either an open version of itself or a closed version which is a toggle button to expand it again.